feat: 일기에 사용자 권한 부여

+++ src/main/java/org/sopt/diary/service/DiaryService.java
@@ -7,6 +7,7 @@ import org.sopt.diary.api.dto.response.DiaryListResponse;
 import org.sopt.diary.api.dto.response.DiaryResponse;
 import org.sopt.diary.domain.Diary;
 import org.sopt.diary.repository.DiaryRepository;
+import org.sopt.member.domain.Member;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -17,8 +18,8 @@ public class DiaryService {
         this.diaryRepository = diaryRepository;
     }
 
-    public DiaryCratedResponse createDiary(DiaryRequest request) {
-        Diary diary = diaryRepository.save(request.toDiary());
+    public DiaryCratedResponse createDiary(Member member, DiaryRequest request) {
+        Diary diary = diaryRepository.save(request.toDiary(member));
         return new DiaryCratedResponse(diary.getId());
     }
 
@@ -33,17 +34,26 @@ public class DiaryService {
         return new DiaryDetailResponse(getDiary(diaryId));
     }
 
-    public void update(long diaryId, DiaryRequest diaryRequest) {
+    public void update(Member member, long diaryId, DiaryRequest diaryRequest) {
         Diary diary = getDiary(diaryId);
+        validateOwner(member, diary);
         diary.update(diaryRequest.title(), diaryRequest.title());
     }
 
-    public void deleteById(long diaryId) {
+    public void deleteById(Member member, long diaryId) {
+        Diary diary = getDiary(diaryId);
+        validateOwner(member, diary);
         diaryRepository.deleteById(diaryId);
     }
 
     private Diary getDiary(long diaryId) {
         return diaryRepository.findById(diaryId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 다이어리 Id입니다."));
+    }
+
+    private void validateOwner(Member member, Diary diary) {
+        if (!diary.getMember().getId().equals(member.getId())) {
+            throw new IllegalArgumentException("해당 일기에 접근할 권한이 없습니다.");
+        }
     }
 }
